Create infosheet.php
Exam information sheets are used to collect information from instructors regarding exam requirements. Instructors can specify whether the exam is a take-home exam, is common with any number of other exams, requires a specific room or type of room, and whether there are religious dates that will affect the exam schedule. (The client-side interaction is managed in standard.js.)